refactor(zustand): rename store files for consistency
