+++ client/src/app/profile/settings/notifications/page.tsx
@@ -4,7 +4,7 @@ import { CardDescription, CardTitle } from "@/components/ui/card";
 import { ApiConfig } from "@/configs/api-configs";
 import { ERRORS } from "@/constants/error-handling";
 import { apiFetch } from "@/lib/apiFetch";
-import { useAlertStore } from "@/store/zustand/alertStore";
+import { useAlertStore } from "@/store/zustand/alert.store";
 import React, { useEffect, useState } from "react";
 
 const $settings = {
