+++ client/src/app/verify/VerifyClient.tsx
@@ -4,7 +4,7 @@ import { CardTitle } from "@/components/ui/card";
 import { ApiConfig } from "@/configs/api-configs";
 import { ERRORS } from "@/constants/error-handling";
 import { apiFetch } from "@/lib/apiFetch";
-import { useAlertStore } from "@/store/zustand/alertStore";
+import { useAlertStore } from "@/store/zustand/alert.store";
 import clsx from "clsx";
 import { useSearchParams } from "next/navigation";
 import React, { useEffect, useState } from "react";
